+++ src/escpos/printer/__init__.py
@@ -1,0 +1,25 @@
+# -*- coding: utf-8 -*-
+"""printer implementations."""
+
+from .lp import LP
+from .dummy import Dummy
+from .file import File
+from .network import Network
+from .serial import Serial
+
+# from .win32raw import Win32Raw
+from .cups import CupsPrinter
+from .usb import Usb
+
+
+__all__ = [
+    "Usb",
+    "File",
+    "Network",
+    "Serial",
+    "LP",
+    "Dummy",
+    "CupsPrinter",
+]
+
+# TODO check which printers are importable and add only those to the namespace
