+++ tox.ini
@@ -22,6 +22,7 @@ deps = jaconv
        pytest-mock
        hypothesis>4
        python-barcode
+       pycups
 commands = pytest
 passenv = ESCPOS_CAPABILITIES_PICKLE_DIR, ESCPOS_CAPABILITIES_FILE, CI, TRAVIS, TRAVIS_*, APPVEYOR, APPVEYOR_*, CODECOV_*
 setenv = PY_IGNORE_IMPORTMISMATCH=1
